Updated API test class
