+++ pom.xml
@@ -87,7 +87,17 @@
 			<artifactId>mockito-core</artifactId>
 			<version>2.21.0</version>
 		</dependency>
-
+		<dependency>
+			<groupId>io.rest-assured</groupId>
+			<artifactId>rest-assured</artifactId>
+			<version>3.3.0</version>
+			<scope>test</scope>
+		</dependency>
+		<dependency>
+			<groupId>com.googlecode.json-simple</groupId>
+			<artifactId>json-simple</artifactId>
+			<version>1.1</version>
+		</dependency>
 		<!-- Validation -->
 		<dependency>
 			<groupId>com.h2database</groupId>
